Change zero division error in graphs
Divide-by-zero errors that occur when graphing functions are now treated as 'holes' in the function, and that point is skipped. Since that could cause points to be wrongly connected, such as with 1/x, curves are changed to be made up of solely points, instead of points connected by lines. This required a change to the number of points and the size of points to make curves seem continuous.